feature: Accept Header based response
Watch for request headers for indication on what type of response is
requested.

Default response is `application/json` with a payload.

If `text/html` is received, and the route handler specifies a `view`
parameter on the data to send `.send({ view: 'home.pug' })`, then
respond in the requested format. Otherwise, send default.